requirements.txt
boto3>=1.28.0
botocore>=1.31.0
pydub>=0.25.1
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
jaconv>=0.3.4
mojimoji>=0.0.12
nltk>=3.8.1
python-dotenv>=1.0.0
google-cloud-speech>=2.25.1
google-cloud-storage>=2.14.0
google-cloud-translate>=3.15.0
numpy>=1.24.0
